python/examples/quick_starters/requirements.txt
composio-crewai
python-dotenv
langchain-openai
